move util functions and correct node replacement
