+++ src/tests/functional/inference_engine/transformations/sr_reshape_1d.cpp
@@ -8,23 +8,8 @@
 #include <ngraph/opsets/opset5.hpp>
 #include <cpp/ie_cnn_network.h>
 
-#include <ngraph/pass/manager.hpp>
-#include "graph_comparator.hpp"
+#include "common_test_utils/ngraph_test_utils.hpp"
 
-namespace {
-
-    void init_unique_names(std::shared_ptr<ngraph::Function> f, const std::shared_ptr<ngraph::pass::UniqueNamesHolder>& unh) {
-        ngraph::pass::Manager manager;
-        manager.register_pass<ngraph::pass::InitUniqueNames>(unh);
-        manager.run_passes(f);
-    }
-
-    void check_unique_names(std::shared_ptr<ngraph::Function> f, const std::shared_ptr<ngraph::pass::UniqueNamesHolder>& unh) {
-        ngraph::pass::Manager manager;
-        manager.register_pass<ngraph::pass::CheckUniqueNames>(unh, true);
-        manager.run_passes(f);
-    }
-} // namespace
 
 TEST(SmartReshapeTests, Reshape1d) {
     std::shared_ptr<ngraph::Function> f(nullptr);
