+++ src/tests/functional/inference_engine/transformations/sr_sub_graph_ops.cpp
@@ -8,25 +8,10 @@
 #include <ngraph/opsets/opset5.hpp>
 #include <cpp/ie_cnn_network.h>
 
-#include <ngraph/pass/manager.hpp>
-#include "graph_comparator.hpp"
+#include "common_test_utils/ngraph_test_utils.hpp"
 
 using namespace ngraph;
 
-namespace {
-
-    void init_unique_names(std::shared_ptr<ngraph::Function> f, const std::shared_ptr<ngraph::pass::UniqueNamesHolder>& unh) {
-        ngraph::pass::Manager manager;
-        manager.register_pass<ngraph::pass::InitUniqueNames>(unh);
-        manager.run_passes(f);
-    }
-
-    void check_unique_names(std::shared_ptr<ngraph::Function> f, const std::shared_ptr<ngraph::pass::UniqueNamesHolder>& unh) {
-        ngraph::pass::Manager manager;
-        manager.register_pass<ngraph::pass::CheckUniqueNames>(unh, true);
-        manager.run_passes(f);
-    }
-} // namespace
 
 TEST(SmartReshapeTests, TensorIteratorStaticParameters) {
     std::shared_ptr<ngraph::Function> f(nullptr);
